+++ app/src/main/java/io/github/mikan/sample/pushnotification/NotificationReceiver.kt
@@ -1,8 +1,8 @@
 package io.github.mikan.sample.pushnotification
 
-import io.github.mikan.pushnotification.NotificationReceiver as BaseNotificationReceiver
+import io.github.mikan.pushnotification.local.NotificationReceiver as BaseNotificationReceiver
 
 class NotificationReceiver : BaseNotificationReceiver() {
-    
+
     override fun getTargetActivity(): Class<*> = MainActivity::class.java
 }
