+++ gradle/libs.versions.toml
@@ -10,9 +10,6 @@ composeBom = "2025.06.00"
 androidxCoreKtx = { group = "androidx.core", name = "core-ktx", version.ref = "coreKtx" }
 androidxAppcompat = { group = "androidx.appcompat", name = "appcompat", version = "1.7.0" }
 material = { group = "com.google.android.material", name = "material", version = "1.12.0" }
-junit = { group = "junit", name = "junit", version = "4.13.2" }
-androidxJunit = { group = "androidx.test.ext", name = "junit", version = "1.2.1" }
-androidxEspressoCore = { group = "androidx.test.espresso", name = "espresso-core", version = "3.6.1" }
 androidxLifecycleRuntimeKtx = { group = "androidx.lifecycle", name = "lifecycle-runtime-ktx", version.ref = "lifecycle" }
 androidxActivityCompose = { group = "androidx.activity", name = "activity-compose", version.ref = "activityCompose" }
 androidxComposeBom = { group = "androidx.compose", name = "compose-bom", version.ref = "composeBom" }
