+++ pushnotification-local/src/main/java/io/github/mikan/pushnotification/local/NotificationScheduler.kt
@@ -1,4 +1,4 @@
-package io.github.mikan.pushnotification
+package io.github.mikan.pushnotification.local
 
 import android.app.AlarmManager
 import android.app.PendingIntent
@@ -7,7 +7,7 @@ import android.content.Intent
 import android.os.Build
 
 object NotificationScheduler {
-    
+
     fun scheduleNotification(
         context: Context,
         notificationId: Int,
@@ -21,16 +21,16 @@ object NotificationScheduler {
             putExtra("title", title)
             putExtra("content", content)
         }
-        
+
         val pendingIntent = PendingIntent.getBroadcast(
             context,
             notificationId,
             intent,
             PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_IMMUTABLE
         )
-        
+
         val alarmManager = context.getSystemService(Context.ALARM_SERVICE) as AlarmManager
-        
+
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
             alarmManager.setExactAndAllowWhileIdle(
                 AlarmManager.RTC_WAKEUP,
@@ -45,7 +45,7 @@ object NotificationScheduler {
             )
         }
     }
-    
+
     fun cancelNotification(context: Context, notificationId: Int, receiverClass: Class<*>) {
         val intent = Intent(context, receiverClass)
         val pendingIntent = PendingIntent.getBroadcast(
@@ -54,11 +54,11 @@ object NotificationScheduler {
             intent,
             PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_IMMUTABLE
         )
-        
+
         val alarmManager = context.getSystemService(Context.ALARM_SERVICE) as AlarmManager
         alarmManager.cancel(pendingIntent)
     }
-    
+
     fun scheduleRepeatingNotification(
         context: Context,
         notificationId: Int,
@@ -73,14 +73,14 @@ object NotificationScheduler {
             putExtra("title", title)
             putExtra("content", content)
         }
-        
+
         val pendingIntent = PendingIntent.getBroadcast(
             context,
             notificationId,
             intent,
             PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_IMMUTABLE
         )
-        
+
         val alarmManager = context.getSystemService(Context.ALARM_SERVICE) as AlarmManager
         alarmManager.setRepeating(
             AlarmManager.RTC_WAKEUP,
